requirements.txt
easydict==1.9
editdistance==0.5.3
h5py==2.8.0
keras
regex==2017.11.9
requests==2.22.0
retrying==1.3.3
six==1.14.0
spacy==2.2.2
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torchtext==0.1.1
tornado==6.0.3
tqdm==4.34.0
yacs==0.1.6
